+++ features/step-definitions/clj-refactor-steps.el
@@ -347,8 +347,8 @@
 
 (Given "The cache of namespace aliases is populated"
        (lambda ()
-         (setq cljr--namespace-aliases-cache
-               (edn-read "{:clj {t (clojure.test)
+         (defun cljr--call-middleware-for-namespace-aliases ()
+           (edn-read "{:clj {t (clojure.test)
 set (clojure.set)
 pprint (clojure.pprint)
 util (refactor-nrepl.util clojure.tools.analyzer.jvm.utils)